feat: only send token to TV
