+++ apps/photos/src/components/Collections/CollectionOptions/AlbumCastDialog.tsx
@@ -12,8 +12,8 @@ import EnteButton from '@ente/shared/components/EnteButton';
 import EnteSpinner from '@ente/shared/components/EnteSpinner';
 import { VerticallyCentered } from '@ente/shared/components/Container';
 import { logError } from '@ente/shared/sentry';
-import { LS_KEYS, getData } from '@ente/shared/storage/localStorage';
 import { Collection } from 'types/collection';
+import { getToken } from '@ente/shared/storage/localStorage/helpers';
 
 interface Props {
     show: boolean;
@@ -72,7 +72,7 @@ export default function AlbumCastDialog(props: Props) {
 
         // ok, they exist. let's give them the good stuff.
         const payload = JSON.stringify({
-            user: JSON.stringify(getData(LS_KEYS.USER)),
+            token: getToken(),
             targetCollectionId: props.currentCollection.id,
             targetCollectionKey: props.currentCollection.key,
         });
